Refactor FileList filter
