+++ assets/src/scripts/outputs-viewer/components/FileList/FileList.jsx
@@ -2,30 +2,31 @@ import React, { useEffect, useRef, useState } from "react";
 import useFileList from "../../hooks/use-file-list";
 import useWindowSize from "../../hooks/use-window-size";
 import useStore from "../../stores/use-store";
+import prettyFileSize from "../../utils/pretty-file-size";
+import classes from "./FileList.module.scss";
 import Filter from "./Filter";
-import List from "./List";
 
 function FileList() {
-  const { listVisible } = useStore();
-  const [listHeight, setListHeight] = useState(0);
   const [files, setFiles] = useState([]);
+  const [listHeight, setListHeight] = useState(0);
 
-  const listEl = useRef(null);
+  const { error, isError, isLoading } = useFileList();
+  const { listVisible } = useStore();
   const windowSize = useWindowSize();
 
-  const { data } = useFileList();
+  const listEl = useRef(null);
 
   useEffect(() => {
     const largeViewport = window.innerWidth > 991;
     const hasScrollbarX =
       listEl.current?.clientWidth < listEl.current?.scrollWidth;
 
-    // Viewport size, minus the Outputs SPA height, minus 30px for spacing
+    // Viewport size, minus the Outputs SPA height, minus 50px for spacing
     // If there are horizontal scrollbars, minus 17px for the scrollbar
     const fileListHeight =
       window.innerHeight -
       document.querySelector("#outputsSPA").getBoundingClientRect().top -
-      30 -
+      50 -
       (hasScrollbarX ? 17 : 0);
 
     if (largeViewport) {
@@ -34,17 +35,55 @@ function FileList() {
     }
 
     return setListHeight(0);
-  }, [data, listVisible, windowSize]);
+  }, [files, windowSize]);
+
+  if (isLoading) {
+    return (
+      <ul className={`${classes.list} list-unstyled card`}>
+        <li>Loading…</li>
+      </ul>
+    );
+  }
+
+  if (isError) {
+    // eslint-disable-next-line no-console
+    console.error(error.message);
+
+    return (
+      <ul className={`${classes.list} list-unstyled card`}>
+        <li>Error: Unable to load files</li>
+      </ul>
+    );
+  }
+
+  const selectFile = ({ e, item }) => {
+    e.preventDefault();
+    return useStore.setState({ file: { ...item } });
+  };
 
   return (
-    <div
-      ref={listEl}
-      className={` card ${listVisible ? "d-block" : "d-none"}`}
-      style={{ height: listHeight || "auto" }}
-    >
-      {data ? <Filter setFiles={setFiles} /> : null}
-      <List files={files} />
-    </div>
+    <>
+      <Filter setFiles={setFiles} />
+      <ul
+        ref={listEl}
+        className={`${classes.list} list-unstyled card ${
+          listVisible ? "d-block" : "d-none"
+        }`}
+        style={{ height: listHeight || "auto" }}
+      >
+        {files.map((item) => (
+          <li key={item.url}>
+            <a
+              href={item.url}
+              onClick={(e) => selectFile({ e, item })}
+              title={`File size: ${prettyFileSize(item.size)}`}
+            >
+              {item.shortName}
+            </a>
+          </li>
+        ))}
+      </ul>
+    </>
   );
 }
